Removed warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@17379 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 572eb031d4754212059b7e1c7b7f85c6c1ffff7b

+++ application/DebsGrandChallenge2013/src/de/uniol/inf/is/odysseus/debs2013/spatial/ShotOnGoalDetection.java
@@ -34,8 +34,8 @@ public class ShotOnGoalDetection extends AbstractFunction<Integer>{
 		if((ts - tsShot) != 0) {
 //			double speedX = ((double) (endX - startX) / (ts - tsShot)) * 1500000000000.0;
 //			double speedY = ((double) (endY - startY) / (ts - tsShot)) * 1500000000000.0;
-			double speedX = (double) (endX - startX);
-			double speedY = (double) (endY - startY);
+			double speedX = endX - startX;
+			double speedY = endY - startY;
 			if(speedX != 0) {
 				double m = speedY / speedX;
 				double b = startY - (m * startX);
